css styling, favicon and footer
Changed filenames in soundclips object within JS file to be all lowercase. Added two favicons into assets directory so damien can choose his preference. Added a footer in the  html file. Changed body resets in the css file as well as styling the buttons into a flex box with wrapping. Body font changed to digital 7 mono to match the theme of the images.